requirements.txt
flask
gunicorn
vertexai
google-generativeai